+++ src/responses/Response.cpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/01 20:44:23 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/20 20:48:00 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/05/21 21:25:14 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -93,10 +93,10 @@ std::string get_file(std::string filename, std::string &mod_date, int *status)
 	std::string		file_content;
 	struct stat file_info;
 	int			file_status = 0;
-	//int			is_dir;
+	int			is_dir;
 
 	file_content = file_reader(filename, &file_status);
-	//is_dir = stat(filename.c_str(), &file_info);
+	is_dir = stat(filename.c_str(), &file_info);
 	if (file_status)
 	{
 		*status = file_status;
@@ -108,7 +108,9 @@ std::string get_file(std::string filename, std::string &mod_date, int *status)
 		return("");
 	}
 	else
+	{
 		mod_date = get_date(file_info.st_mtime, true);
+	}
 	*status = 0;
 	return(file_content);
 }
@@ -135,12 +137,19 @@ void Response::file_status_custom_error(int file_status)
 		else if (!request.get_dir_params()->second.index.empty())
 		{
 			std::string file_to_get = request.get_path_abs();
+			
+			file_to_get.append("/");
 			file_to_get.append(request.get_dir_params()->second.index);
-			get_file(request.get_dir_params()->second.index, mod_date, &file_read_status);
+			get_file(file_to_get, mod_date, &file_read_status);
 			if (file_read_status == ENOENT)								//?	File Not found
 				status_code = 404;
-			else if (file_status == EACCES || file_status == EISDIR)	//?	Permission denied or is a directory
+			else if (file_read_status == EACCES || file_read_status == EISDIR)	//?	Permission denied or is a directory
 				status_code = 500;
+			else if (!file_read_status)
+			{
+				status_code = 200;
+				
+			}
 			else
 				status_code = 501;
 		}
@@ -160,6 +169,7 @@ void Response::file_status_custom_error(int file_status)
 
 void Response::return_content()
 {
+	std::cout << "";					//?	Fixes a Sanitizer error somehow.
 	std::string mod_date;
 	int status = 0;
 	std::string get_body = get_file(request.get_path_abs(), mod_date, &status);
@@ -175,7 +185,13 @@ void Response::return_content()
 	else if (status)
 	{
 		file_status_custom_error(status);
-		head_params["Content-Type"] = get_filetype.get("html");
+		if (status_code == 200)
+		{
+			// ! Return the file type and mod date of the file inside the directory.
+			body = get_body;
+		}
+		else
+			head_params["Content-Type"] = get_filetype.get("html");
 	}
 	else
 	{