+++ src/responses/Response.hpp
@@ -6,7 +6,7 @@
 /*   By: jvacaris <jvacaris@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/01 20:44:32 by jvacaris          #+#    #+#             */
-/*   Updated: 2023/05/20 20:25:45 by jvacaris         ###   ########.fr       */
+/*   Updated: 2023/05/21 19:14:46 by jvacaris         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,6 @@ class Response
 	public:
 		Response(const Request &_request);
 		void return_content();
-		std::string reply();
 		std::string generate_response();
 		void file_status_custom_error(int file_status);
 		void return_error_message(int error_code);